[성진] 파일 BLOB으로 바꿀때 null값 넘어가는것 수정

+++ src/main/java/com/kh/billida/rentalHistory/model/dto/Rental.java
@@ -7,12 +7,12 @@ import lombok.Data;
 @Data
 public class Rental {
 	
-	private long historyIndex;
+	private Long historyIndex;
 	private Long lockerId;
 	private String userCode;
 	private Date rentStart;
 	private Date rentEnd;
-	private int rentCost;
+	private Long rentCost;
 	private String rentPic;
 
 }
